Move Model out of SheepModel.kt.
